chore: move persistent session context manager into shared file
